trainer/src/requirements.txt
pycocotools==2.0.0
coremltools==2.0
tfcoreml==0.3.0
tensorflowjs==0.8.0
tensorflow-hub==0.3.0
h5py==2.8.0